2.11. "Fixed" an FC when viewing certain contacts. Quote from http://code.google.com/p/cyanogenmod/issues/detail?id=890:
"As I expected, it's because one of your IM addresses is "GTALK: xxx@googlemail.com"
instead of simply "xxx@googlemail.com". The GTALK prefix is causing the issue. Did
you use any other contacts application?

I've fixed it by catching the exception, since "GTALK: xxx@googlemail.com" is NOT
parsable as an IM address, thus the protocol type can't be detected. Will upload 2.11
in a while, hopefully in time for cyanogen to merge in.

Take note, however, this isn't really an issue with Contacts.apk, since if an app is
saving the Contact info correctly, it should not have the "GTALK: " or any other sort
of prefix."

Not really mine or AOSP's code's problem, but well :).