ContactList: Prevent Dialer from loading when Contacts app is launched.
This change prevents Dialer from being setup as the current tab when
Contacts is launched. This improves performance by preventing the
TwelveKeyDialer class from getting loaded when Contacts List is viewed.
When the Phone app is launched, the original behavior is unchanged and
the Dialer is set up as the current tab.

Change-Id: I0794dc860d55565f3a33ea65484e5ae8a26d89bd
CRs-Fixed: 228919