Update patch set 1
Patch Set 1:

I made this change because a person with small vision problems  that I know can not read the numbers in the phone book because they are too small.

Also in my opinion they can be difficult to read, especially for older people.

The simple change I made allows you to select the text size of the "data" row, to understand better look at the screenshots:

http://www.anddev.it/adipic/images/fontsmallvlv.png
http://www.anddev.it/adipic/images/fontsizeseleogo.png
http://www.anddev.it/adipic/images/fontbigasa.png

Patch-set: 1