Update patch set 1
Patch Set 1:

Thanks, Keyan - the only drawable that was deleted (the rest were shuffled around because there are h/mdpi versions) was ic_launcher_smsmms, which seemed to be just an ad-hoc grayscale scaled copy of the SMS launcher icon. I hope nobody is using that! ;-) But, of course, it would be easy enough to leave a copy of that around, if needed.

Patch-set: 1