Update patch set 1
Patch Set 1:

Meaning it makes absoulutley no difference if it is spelled 'dialed' (US) or 'dialled' (World)...everyone will understand what it means either way.

While I appreciate everyone picking up the code and hacking away at it, I can't understand arguments like this. If the majority speaks and says they like things a certain way...this is the way it should be..wrong or right. Or at least I believe the majority rules in a community driven project.

Patch-set: 1