Update patch set 1
Patch Set 1:

T. Turner: While a community does rule in a community driven project, this change is as trivial as it is wrong.  There is an en_US locale and an en_GB locale.  The en_US caters to the United States and is also the default locale of CyanogenMod.  The en_GB locale is used for Great Britain.  Rather than reverting this change, the proper pathway would be to "translate" these values for the proper locale.  Dialed is the correct spelling in en_US.

Patch-set: 1