Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

(1 inline comment)

Contacts has a preference to enable sensor rotation.  The dialer screen will crash when it is turned to landscape mode (the dialer is designed to return null in landscape so it will not be drawn).

This patch removes the call to setupKeypad causing the crash.  In this context it is being used to redraw the keypad if the voicemail number has changed.  This shouldn't happen very often, so may not be necessary. 

In landscape mode, numbers can still be dialed with a popover dialpad even though the original dialpad is not drawn.

Patch-set: 1