Contact shortcut fc fix
Currently, direct contact details cannot be launched via shortcut. Only via quickShortcut.
However cmparts does not make use of this showQuickContact and for valid reasons. A long press of a button need not
launch the quick bar at a random position, rather should take you to that contacts details.

Currently setting a contact shortcut via cmparts for a button results in a force close (except for long press home, since it was patched),
because sourceBounds is not set in the intent. In aosp, if sourcebounds is not found, upon launch it will force close
complaining "Missing anchor rectangle" because the show attribute for QuickContactWindow has a :

if (anchor == null) {
      throw new IllegalArgumentException("Missing anchor rectangle");
}

This is aosp behaviour, not something implemented in cm.

As a work around this patch checks prior to launching that method whether the target sourceBounds attribute is null,
if it is then instead of proceeding to end up with a force close, it launches the contact details for that contact with a helpful log message.

Change-Id: Idb825ac935d9e6a9b30618e64081cc6a8c1ca008