Update patch set 2
Patch Set 2:

A few more notes : 

Yes, this does change standard behaviour. 

In aosp, ud get a force close if you tried to launch QUICK_CONTACT without a sourceBounds attribute, or tried to use showQuickContact with null as the second attribute.

With this, instead of a force close, u'll be taken to the contacts details activity and in the log a helpful message will let you know why that happened.

Patch-set: 2