Update patch set 2
Patch Set 2:

An alternative would be replacing the QUICK_CONTACT intent string with a VIEW one in CMParts. It's done already there, but at a lot of different places and not consequently.

We can either do the change proposed here or clean up CMParts and fix up the intent there - I'm really unsure which fix is better.

Patch-set: 2