Update patch set 7
Patch Set 7:

I'm worried about how this will scale. Has anyone tested it with thousands of contacts? When I was prototyping this, I used glob queries to the provider to only fetch data that matched. Is it a good idea to fetch all contacts into a list and filter them like this?

Patch-set: 7