Update patch set 7
Patch Set 7:

Steve,

We first had it that way, where we'd do live queries. We were noticing a lag though, very small (~0.5-1s) but still noticeable. So then we switched to this method as a comparison and the results were noticeably faster.

However, since then we've done a lot of *optimization's* to the query/sorting/filtering, for one pawitp got rid of all regex stuff which seemed to be the culprit for the slow load times...etc. So now im keen to go back to the old method and try it out again with all these refined changes/tweaks. I have the day off tomorrow, so let me take a stab at it and i'll report back the results.

Patch-set: 7