Update patch set 8
Patch Set 8: I would prefer that you didn't submit this

I don't think it is possible to do database searching without modifying the database and other parts of the contact application to store the numerical equivalent of the contact. The only problem with loading lots of contacts would be the load time.

-1 because there's a bug where the database is always loaded regardless of whether T9 is enabled or not.

Patch-set: 8