Update patch set 2
Patch Set 2:

I like it!  Couple of issues though:
When you first open the dialer, before pressing any numbers, if you touch the screen where a name would appear (before there is one to press) it crashes.
It's fine if you have entered and then deleted a number.  Presumably it starts up with something null which shouldn't be?
On a 480x800 screen (ZTE Blade) the tops of the numbers you are dialling are cut off

Patch-set: 2