Update patch set 2
Patch Set 2:

Arne,

Once again, ive forgotten about ldpi users :(
I'll upload an updated overlay layout for contacts by evening if no one else has. Sorry, for all the headaches this might've caused.

Mark,

I can't reproduce that on the ics t9, can't say about gingerbread one because im no longer on it. The view is suppose to be invisible when no digits are entered so not sure how its allowing you to click it. But a fix is quite trivial. As for the cutoff issue, danny has uploaded a patch to fix it.

Patch-set: 2