Update patch set 5
Patch Set 5:

I can confirm the fixes detailed for patchset 5 have worked for me.
The keyboard no longer pops up for the number edit box.  And after selecting a number from the list, my typing cursor is moved to the end of the digits.  If I don't drop the list down however and click the first choice offered, the cursor isn't placed in the phone number box.  Not sure if it should for consistency?

I've also just noticed in patchset 4 and 5 (not sure if I had this problem before this 'T9 Dialer - Fixes' patch) all of the entries in the drop down are duplicated 3 times.

Patch-set: 5