Update patch set 6
Patch Set 6:

The thing with the cursor has always been like this (xml sets cursorVisible="false", and on the first click in the editbox it's set to true).

About those multiple entries: It's fine here, no duplicates. I didn't change the code that creates the list, either, so this is a weird problem. Can anyone confirm?

Patch-set: 6