Update patch set 6
Patch Set 6:

The problem appears to be in my data.  I downloaded sqlite editor, loaded up the contacts2.db and opened view_v1_people and everything is in there 3 times.  Weird because before reporting the issue, in case it was a data problem, I did clear data on contacts and resync it with google.  So it seems my contacts are screwed up in gmail and it's filtering out the dupes (as is the android contacts app)

I'll have to have a look around and see if I can fix duplicate contacts somehow!

Patch-set: 6