Update patch set 1
Patch Set 1:

This is a quickfix in order to keep ContactItem static. 

Actually TwelveKeyDialer does instantiate a ContactItem() as a dummy for "no results". 

BitmapContactItem needs access to the outer mContext in getPhoto(), so it may not be static (or would need its own mContext private field).

In order to keep the change small, I decided for this solution.

Patch-set: 1