Update patch set 2
Patch Set 2:

Got me, I did not test the performance change from array to map on its own, only the complete change -- which acts more snappy and less OOM than the original. 

Thanks for sharing your benchmark code, so I tested something else and now have a modified patch set that actually has a faster name to number conversion as well. 

But in the end (after playing around with the benchmark code) I think the conversion part is not the important point here, but the image reading is.

Regards, Markus

Patch-set: 2