Update patch set 1
Patch Set 1:

My testing on I9100G shows that the search only works on the first letter of the Pinyin for a character. Typing more than the first letter of the Pinyin for a character would stop the search from working. eg. http://pastebin.com/X38YquHy (Gerrit can't display Chinese properly)

Also with this, T9 would not search for contacts with English names while the language is set to Chinese. 

After setting the language to English, it would continue to search by Chinese Pinyin until you clear the data to the Contacts app, which at that point, T9 would search for contacts in English once again.

Patch-set: 1