Update patch set 2
Patch Set 2:

The problem of the search engine not switching back is still present. This can be seen when you start off with English, and when you switch the language to Chinese, it will not search in Pinyin.

If you start off with Chinese and you switch to English, Chinese would continue to be searched, but that's not as big of a problem since English support is still there.

Patch-set: 2