Update patch set 2
Patch Set 2:

Danny,

Ah must've forgotten to remove that.
If you check onResume of that activity, you can see its set based on the shared preference value.

And yes, perhaps get a user with quite a few contacts who actually gets a loading bar or something and see how changing orientation effects their experience

As for getColumnByIndex, i just saw it being used a lot in the package as well so thought id mention it. But your call :)

Patch-set: 2