Update patch set 3
Patch Set 3:

You move the load from onResume to onStart in patch set 3 to avoid load T9 every time for performance.But the event onStart only call one time on view create. It's not same as onResume will call every time when the activity showwing. So I thing switch the locale  will has problem. The constacts change has same problem also. I am apt to add linstener for the two cases, and adjust in onResume to check the need load again or not. It can get good performance with right result that we are expect. Is there any other case we lost? I am not sure.

Patch-set: 3