Update patch set 3
Patch Set 3:

Please forget my comment on Patch Set 3, I miss the Activity.onStart and Fragment.onStart. Patch Set 3 onStart will call when the Fragment visiable. But need we load the T9 every time? This is poor performance for the case if there are a lot of constacts. Still not see and used the patch set 4 code. Same as you add cache for it. If patch set 4 can only load the T9 in three cases, it's good enough to me.
1. The first time load.
2. When user switch locale.
3. When the user add/delete/update the constacts in other application.

Patch-set: 3