Update patch set 12
Patch Set 12: (1 inline comment)

There's no real need for the abbreviation (in the German translation, I use 'Stunden', 'Minuten' and 'Sekunden'), I just followed the callDetailsDurationFormat string. In fact, I thought about completely replacing callDetailsDurationFormat with this new format as it does some things quite a bit better (it e.g. shows '1 min' instead of '1 mins 0 secs'), but I didn't do it because it would mean losing the already existing translations for it. Parts of the translations could be importec from DeskClock (which is where I borrowed the new logic from), but it's day/hour/min instead of hour/min/sec there.

Opinions on this?

Patch-set: 12