Update patch set 18
Patch Set 18:

I'am not running the newest Patch Set and when no one else hase complaind about this issues I think we are ready to go. When I have the same problem after this is merged I will tell you and then it's better to make a new commit to fix it, but O don't hope so :)

Merge this awesome stuff!

Patch-set: 18