Update patch set 1
Patch Set 1:

If it's the same as the default translations, I think we don't need to add it. It might be helpful to add a feature to Translation Tool to set a string untranslatable on a per language basis, such as the 'no translation found' line you can see in some languages.

Patch-set: 1