Contact: Phone crash when save a contact after switch the account
The state of the account switcher is incorrect after end a MT call,
the contact has been saved and the action is "EDIT" after end a MT
call, so user should can't switch account now.

Add a judgment to update the value of the variable that indicates
whether there is a new contact, the program will disable the account
switcher according to this value.

CRs-Fixed: 528413

Change-Id: Iaf3dbddc4803f7120c1a7613e42997cb72684387