Contacts: After update the displayed members, group can match it again.
Group use a list which contain the displayed members' contact id to avoid
add same member to the group.
After we pause the activity and back again, it will load again. But when
load finished, it will only add the members which existed to the list,
and missed the members which will to be added. So the members which will
to be added can be matched again.

When load finished, the displayed members contain the members which
existed and  will to be added. So we should add the displayed members to
the list.

CRs-Fixed: 556582

Change-Id: I43baa9e44aca0cc00942f0f6edb400faf49af891