Update patch set 3
Patch Set 3:

Small note(s):

On this patch, explicitly clicking on the checkbox next to the contact icon now doesn't add the contact to bundle to be returned.

Should we switch to a CAB outright when we start selecting items as to follow design guidelines a bit more closely? I rushed the UI change originally to just add the "done" item to the action bar when anything is selected.

Patch-set: 3