Update patch set 3
Patch Set 3:

I should probably make the checkbox non-clickable.

About the CAB, I thought about leveraging the listview multi choice action mode. The problemn with that is synchronizing the search views of the normal action bar and the CAB. If there's a solution to that, I'm all for using a CAB.

Patch-set: 3