Update patch set 1
Patch Set 1:

This is a consequence of a recent change that modified how the exporting of contacts to storage work. Previously, when you selected to export contacts to storage, it automatically went ahead and exported all of them. Apparently, some recent change made it so that you are taken to a contact picker screen first. From what I can tell, the MAX_CONTACTS_NUM_TO_SELECT_ONCE is not specific to contact export, but to mass contact operations in general (adding to groups, etc). Hence I am not sure what implications modifying the value will have, nor do I know why it is needed in the first place. 

Again, setting it to 1000 doesnt really make much more sense than 500. Android is obviously more than capable of doing mass contact operations on an very large number of contacts.

Patch-set: 1