Contacts: Choose dialog show again after select "Always"
There is a "NonPhoneActivity" in contact package
has the same intent-filter with "CallLogActivity"
but it's priority is "-1" which other activity's
is "0", so the result is "false" always when
compared by priority. This cause choose dialog
show again after select "Always".

As the intent-filter of NonPhoneActivity is not
use, so delete it.

CRs-Fixed: 736376

Change-Id: I80f2217e4727d267a4932ebdde1412699f2b0c2e