Contacts: Can not remove photo after rotating the phone
If we go to Mms and rotate phone after we've setted a photo
for the contact, Contacts will save the photo to db and the
state will be inited again, ValuesDelta's status will be
lost. As a result, we can not delete the photo when we come
back to Contacts.

Add configChanges in manifest file to prevent being destoried
when rotating phone.

CRs-Fixed: 748891

Change-Id: I46d7347816a2e9b2951b62a00f35dd97824a3fe6