Contacts: Can not search out the contact after removed it from group
Before showing the results that match the filter string, Contacts
will remove the results that has added to the group. The root cause
of this issue is that the array list that holding the existing
group members is not correct.

Add the contact id to the list only when the contact id is not added
to the list yet. Besides, we should not show the raw contacts have
been deleted.

CRs-Fixed: 753166

Change-Id: I4fdbea3bdacc874108c5c4a51165f0c6525a5e66