Contacts: Can not set photo when importing contacts
Importing contacts will cause the background data changement,
if we we are editing contact when background data is changing,
we do not want to change our data in the UI. However, the
status will be changed to EDITING in the onLoadFinished which
is not proper since our data in the UI isn't changed.

Do not change the status in onLoadFinished if it is caused
by background data changement.

CRs-Fixed: 748192

Change-Id: I5d99d3f25158c5989d72316540aef76ff00a8934