Contacts: Searching function in group editting interface is not powerfull
The default threshold of AutoCompleteTextView is 2, so we
can not search anything with one letter. The logic we used
to search contacts in group editing interface is querying
all the contacts' display name that match the given search
query, it does not support searching contacts' other info.

Change the threshold to 1. Use contacts' filter uri instead
of rawcontacts' content uri to implement a much more powerfull
search function.

CRs-Fixed: 761634

Change-Id: I9858136175cdd35f036bdaa4dc3c40df90aa7deb