Update patch set 1
Patch Set 1:

Only other thing I would wonder if it's some piece of data that some people have for their contacts.

The function "dataItemToEntry" is responsible for loading the data entries into the item that I've patched to check for null.  Maybe something within there is causing it to come back empty.  Just not sure what.

When I first tried this, I half expected an empty screen for contact details.  Assuming the list was null, how could there be information shown.  But that's not the case, haven't seen any negative from this.

Patch-set: 1