Always use contacts sms icon for quick contact card
There is logic in the quick contact card that attempts to fetch a drawable
named ic_text_holo_light from the package that is the source for the contact
information being displayed.  This will never work correctly because we are
taking the ID for that resource as defined in contacts app and searching for
it in another package (which will have different id mappings).  In the worst
case scenario, the package will have a completely unrelated drawable that maps
to the same ID, and we will attempt to display it as the sms icon.

To get around this, always use the sms icon defined in contacts

Change-Id: I4ab61ed73a769d618a75071a175f142b4d248486