Keep a local instance of mContactData when copying to SIM card.
Copying to SIM card can take a long time while it reads data about
available space on the SIM. If the user leaves the activity,
mContactData will be set to null and copyToCard will throw an NPE trying
to call methods on it.

Keep a local instance of this variable that the
thread can reference later.

Change-Id: Ie5187372b6fb2b23b00e266a175c91b8051e651e