Contacts: Fixed "%s" appearing in contact options in some languages
In the contacts-app, when looking at a contact, there is a entry in the options called "Copy to SIM" (or similar) in most cases.
The text in the translation file for this is "Copy to <xliff:g id="destination">%s</xliff:g>" (%s at the end), while the text is "Nach <xliff:g id="destination">%s</xliff:g> kopieren" (%s in the middle) in the German translation for example.
The problem is that the destination (e.g. "SIM") is added using the plus operator instead of replacing the %s.
This does not cause any trouble if the %s stands at the end, such as in English, but in German it causes "Nach %s kopierenSIM" instead of "Nach SIM kopieren".
Adding "SIM" as second parameter in getString() instead of using "+" should replace %s by "SIM" and fix the problem.

Change-Id: I81d3d1b366e7bbff3e331af67295c53100f5840a