Contacts : only refresh ENABLED plugins onResume
Since we only request refreshDynamicItems if there are ENABLED
plugins, there would not be a callback that calls updatePlugin.
In that case we'll manually call updatePlugin to manually check
if any plugin related UI components need to be updated (eg.
the current UI may show a plugin that's already gone into
DISABLED/HIDDEN state)

CD-616

Change-Id: I469960212e67844a63661e855daa7097ac186a05