Use the explicit spam flag on Contact for identifying spammy contacts
We were using the spam count as an indicator for determing whether a
contact is spam. Changing this behavior to rely on a caller-info provider
to signal that a contact is spam, rather than determining that ourselves.
A caller-info provider would have the proper domain knowledge to make
this determination, so deferring to them is worthwhile.

Change-Id: If548926b7269e51c0805f5d1ef73edcffbd9e298
Issue-Id: OPO-736
(cherry picked from commit 66374db2f803b0f399473b440022b62eb4db8e9c)