Update patch set 1
Patch Set 1:

There's some issues with this.

If your phone number is stored on the SIM card, the current user will be listed as the number 2 in settings, instead of Owner.

If your phone number is not stored on the SIM card, current user in settings is correctly listed as Owner. But if you then open contacts app, edits your profile and adds your phone number, then current user will show as 2.

Patch-set: 1
CC: Arne Coucheron <444@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>