Update patch set 1
Patch Set 1:

Stored on the sim means the SIM number, which you can set with for example this: https://review.lineageos.org/#/c/LineageOS/android_packages_services_Telephony/+/216722/

This patch are reading that number from the SIM, and prefills the phone number in your profile. Normally most SIM cards doesn't have this number preprogrammed in the SIM. If you check in SIM status what's listed under "My phone number", you'll see if that's the case. If it says Unknown, the first user will be Owner. If it lists a phone number, the first user will be 2.

If you have an unknown phone number listed, you can trigger the bug by editing your personal info in Contacts, and fill in your phone number.

Patch-set: 1