Revert "Automatically set SIM number to my profile"
This commit introduces an issue within the SIM number and users handling.

In Settings / System / About / Status / SIM Status, the field "My phone number"
will either be "Unknown" or the real number, depending on what is stored on the
SIM card.

If it is the real number, the "Current user" in Settings dashboard will be
listed as the number "2" instead of "Owner".

If it is "Unknown", the "Current user" is correctly shown as "Owner".
But if you open the AOSP Contacts app, edit your profile and add your phone
number, the current user will be shown as "2".

Correct: http://i65.tinypic.com/33vnofa.jpg
Bug: http://i63.tinypic.com/1zwoh2a.jpg

This reverts commit 7f3c3a32b5d0dd02605c761eb0982dee1878ed92.

Change-Id: Idcab75bc99c5f1d47cd8645417e7d3ba27659a5c